Platform code cleanup, const correctness, int -> bool
Cleaned up platform specific code:
Removed ULONG, replaced with unsigned int
Hash no returns a CRC32_t
Made some functions const correct
Made CBaseEntity::GetMonsterPointer static

+++ dlls/client.cpp
@@ -44,10 +44,10 @@
 #include <ctype.h>
 #endif
 
-extern DLL_GLOBAL ULONG		g_ulModelIndexPlayer;
-extern DLL_GLOBAL bool		g_fGameOver;
-extern DLL_GLOBAL int		g_iSkillLevel;
-extern DLL_GLOBAL ULONG		g_ulFrameCount;
+extern DLL_GLOBAL unsigned int	g_ulModelIndexPlayer;
+extern DLL_GLOBAL bool			g_fGameOver;
+extern DLL_GLOBAL int			g_iSkillLevel;
+extern DLL_GLOBAL unsigned int	g_ulFrameCount;
 
 extern void CopyToBodyQue(entvars_t* pev);
 extern int giPrecacheGrunt;
